Split tests based on what they check
