+++ package.json
@@ -1,19 +1,18 @@
 {
   "devDependencies": {
     "apache-server-configs": "2.11.0",
-    "archiver": "^0.11.0",
+    "archiver": "^0.12.0",
     "del": "^0.1.3",
     "glob": "^4.0.6",
     "gulp": "^3.8.9",
     "gulp-header": "^1.2.2",
     "gulp-jscs": "^1.2.1",
-    "gulp-jshint": "^1.8.5",
+    "gulp-jshint": "^1.8.6",
     "gulp-load-plugins": "^0.7.0",
     "gulp-rename": "^1.2.0",
     "gulp-replace": "^0.4.0",
     "jquery": "1.11.1",
     "jshint-stylish": "^1.0.0",
-    "lodash": "^2.4.1",
     "mocha": "^2.0.1",
     "normalize.css": "3.0.2",
     "run-sequence": "^1.0.1"
